filter-bot/requirements.txt
aiogram>=3.0.0
redis>=4.2.0
python-dotenv
loguru 